🎉 Moves Interfaces to their own project
