+++ Configuration.Persistence.Contracts/Behavioral/IAuditable.cs
@@ -1,14 +1,10 @@
-namespace Kritikos.Configuration.Persistence.Abstractions
+namespace Kritikos.Configuration.Persistence.Contracts.Behavioral
 {
-  using Kritikos.Configuration.Persistence.Interceptors;
-
   /// <summary>
   /// Exposes barebones auditing functionality on a multi-user system.
   /// </summary>
   /// <typeparam name="T">Type of auditor identifying field.</typeparam>
-  /// <remarks>
-  /// Fields are set automatically via <see cref="AuditSaveChangesInterceptor{T}"/>.
-  /// </remarks>
+  /// <remarks>Fields should be handled by interceptors.</remarks>
   public interface IAuditable<T>
   {
     T CreatedBy { get; set; }
