+++ Configuration.Persistence.Contracts/Behavioral/ITimestamped.cs
@@ -1,15 +1,11 @@
-namespace Kritikos.Configuration.Persistence.Abstractions
+namespace Kritikos.Configuration.Persistence.Contracts.Behavioral
 {
   using System;
-
-  using Kritikos.Configuration.Persistence.Interceptors;
 
   /// <summary>
   /// Exposes timestamping functionality for barebones persistence auditing.
   /// </summary>
-  /// <remarks>
-  /// Fields are set automatically via <see cref="AuditSaveChangesInterceptor{T}"/>.
-  /// </remarks>
+  /// <remarks>Fields should be handled by interceptors.</remarks>
   public interface ITimestamped
   {
     DateTimeOffset CreatedAt { get; set; }
